nothing

+++ core/src/com/bitfire/postprocessing/effects/LensFlare2.java
@@ -217,7 +217,7 @@ public final class LensFlare2 extends PostProcessorEffect {
 	}
 
 	public int getGhosts () {
-		return (int)lens.getGhosts();
+		return lens.getGhosts();
 	}
 
 	public boolean isBlendingEnabled () {
